pyproject.toml
[project]
name = "pycommit-ai"
version = "0.1.0"
description = "AI-generated Git commits"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "google-genai>=1.64.0",
    "groq>=1.0.0",
    "inquirerpy>=0.3.4",
    "openai>=2.21.0",
    "rich>=14.3.3",
]

[project.scripts]
pycommit-ai = "pycommit_ai.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pycommit_ai"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
